handle single instance of google oauth
